requirements.txt
Django==5.2.4
djangorestframework
djangorestframework-simplejwt
gunicorn
psycopg2-binary
python-dotenv
whitenoise